remove old shit, scream test refactor (meaning there's probably still bugs around lol)
