+++ src/components/layers/hermes/hermes.css
@@ -132,7 +132,7 @@
 	padding-bottom: 5px;
 	padding-right: 6px;
 	gap: 5px;
-	height: 400px;
+	height: 300px;
 	overflow: auto;
 	overflow-x: hidden;
 	/* overflow: hidden; */
