--- src/components/ui/event-log/EventLog.tsx
@@ -1,57 +1,0 @@
-import CornerRect from '@/components/util/corner-rect/CornerRect';
-import tr from '@/assets/ui/corners/d1/tr.png'
-import br from '@/assets/ui/corners/d1/br.png'
-import tl from '@/assets/ui/corners/special/ovl_bot.png'
-import { createEffect, createSignal, For, on } from "solid-js";
-import './event-log.css'
-
-const [logMessages, setLogMessages] = createSignal<{ id: number, text: string, color: string }[]>([]);
-
-/**
- * Append a message to the "EventLog" which is the small text box at the bottom of the screen.
- * @param msg Message to append
- */
-export const addLogMessage = (text: string, color = "#aaa") => {
-  // Use date to force uniqueness.
-  setLogMessages((prev) => [...prev.slice(-15), { id: Date.now(), text, color }]);
-};
-
-// Little greeting message for demo. Remove later.
-setTimeout(() => {
-  addLogMessage('Welcome to daemon.garden', 'teal')
-}, 50)
-
-/**
- * Textbox at the bottom of the Main UI, under the scene. Used to convey game events or other short bits of info.
- * Text can be appended to this from anywhere using the addLogMessage method.
- */
-export default function EventLog() {
-  let containerRef: HTMLParagraphElement | undefined;
-
-  createEffect(
-    // Trigger scroll down when messages change (append)
-    on(logMessages, () => {
-      if(containerRef) {
-        containerRef.scrollTo({ top: containerRef.scrollHeight, behavior: "smooth" });
-      };
-    })
-  )
-
-  // Scrolljacking since default scroll distance on every tested web browser was too large, skipping over text.
-  const handleWheel = (e: WheelEvent) => {
-    if(containerRef) {
-      e.preventDefault();
-      containerRef.scrollTop += e.deltaY * 0.3;
-    }
-  }
-
-    return (
-      <CornerRect borderSize={3} borderType="double white" corners={[tl, tr, undefined, br]} style={{width: 'inherit', height: 'inherit'}} id="event-log">
-        <p class="inner" ref={containerRef} onwheel={handleWheel}>
-          <For each={logMessages()}>
-            {(msg) => <p class="event-message" style={{color: msg.color}}>{msg.text}</p>}
-          </For>
-        </p>
-      </CornerRect>
-    )
-}
