--- src/components/ui/scene-container/SceneContainer.tsx
@@ -1,62 +1,0 @@
-import {createSignal, Show, Suspense} from "solid-js";
-import { scenes } from '@/scenes/sceneRegistry';
-import { Dynamic } from "solid-js/web";
-import { INITIAL_SCENE, SCENE_DIMENSIONS } from "@/config";
-import { currentInteractionMode, setCurrentInteractionMode } from '../interaction/InteractionModePicker';
-import CornerRect from "@/components/util/corner-rect/CornerRect";
-import { DialogueService } from "@/core/dialogue/dialogueService";
-
-import './scene-container.css'
-import tl from '@/assets/ui/corners/s5/tl.png'
-import tr from '@/assets/ui/corners/s5/tr.png'
-import br from '@/assets/ui/corners/s5/br.png'
-import bl from '@/assets/ui/corners/special/ovl_top.png'
-
-/**
- * Use this atom to change/view the active rendered scene. 
- * Changing the scene will completely unmount the previous scene and 
- * immediately load the new one.
- */
-export const [currentScene, setCurrentScene] = createSignal(INITIAL_SCENE);
-
-/**
- * Scene container is responsible for the dynamic lazy loading of scenes from the scene registry. It listens to the public
- * currentScene atom to load and switch between the active scenes.
- * 
- * To change a scene, import setCurrentScene atom and call it by name. (You can also get the active scene by import.)
- * 
- * To register a scene, reference sceneRegistry.
- * @returns 
- */
-export default function SceneContainer() {
-
-    // Used to cycle interaction mode if you right click the scene.
-    function cycleInteractionMode() {
-        setCurrentInteractionMode(
-            (currentInteractionMode() + 1) % 3
-        )
-    }
-
-    return (
-        <CornerRect 
-            onContextMenu={cycleInteractionMode} 
-            borderSize={1} 
-            borderType="solid white" 
-            corners={[tl, tr, bl, br]} 
-            id="scene-container" 
-            style={{
-                width: `${SCENE_DIMENSIONS.width + 2}px`, 
-                height: `${SCENE_DIMENSIONS.height + 2}px`,
-                //pointerEvents: DialogueService.activeDialogue()? "none" : "auto" // doesn't work :(
-            }} // need +2 for borders. Change with border size thanks.
-        >
-            <Suspense fallback={<p>Loading scene...</p>}>
-                <Dynamic component={scenes[currentScene()]} />
-            </Suspense>
-        {/* We always have an overlay on dialogue, it works as an easy blocker for mouse events AND we can show artwork when we need :) */}
-        <Show when={DialogueService.currentDialogueOverlay()}>
-            <div id="dialogue-overlay" style={{background: `url(${DialogueService.currentDialogueOverlay()})`}}></div>
-        </Show>
-        </CornerRect>
-    )
-}
