--- src/components/ui/scene-container/scene-container.css
@@ -1,16 +1,0 @@
-#dialogue-overlay {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-    background-color: #000000aa;
-    opacity: 0;
-    animation: forwards 0.5s fadeIn;
-}
-
-@keyframes fadeIn {
-    to {
-        opacity: 1;
-    }
-}
