--- src/components/ui/sidebar/Sidebar.tsx
@@ -1,53 +1,0 @@
-import { scenes } from "@/scenes/sceneRegistry";
-import { For } from "solid-js";
-import { currentScene, setCurrentScene } from "../scene-container/SceneContainer";
-import CornerRect from "@/components/util/corner-rect/CornerRect";
-import tl from '@/assets/ui/corners/s5/tl.png'
-import tr from '@/assets/ui/corners/s5/tr.png'
-import bl from '@/assets/ui/corners/s5/bl.png'
-import br from '@/assets/ui/corners/s5/br.png'
-import { pushUILayer } from "@/core/ui/UILayerStore";
-import { TestComp } from "@/components/layers/UILayerHost";
-import { startBattle } from "@/core/battle/battleManager";
-import { OPPONENT_PANOPTES } from "@/data/battles/panoptes";
-import { OPPONENT_DEBUG_ANGEL } from "@/data/battles/debugangel";
-
-export default function Sidebar() {
-
-    return (
-        <CornerRect id="sidebar" borderSize={1} borderType="solid white" corners={[tl, tr, bl, undefined]} style={{width: 'inherit', height: 'inherit'}}>
-            <CornerRect id="debug-menu" borderSize={1} borderType="solid white" corners={[tl, tr, bl, br]} style={{margin: '4px', height: 'fit-content', padding: '5px'}}>
-                <h3>DEBUG MENU</h3>
-                <h4>Change scene:</h4>
-                <p>Current scene: {currentScene()}</p>
-                <For each={Object.keys(scenes)}>
-                    {(scene) => (
-                        <button onClick={() => setCurrentScene(scene)}>
-                            {scene}
-                        </button>
-                    )}
-                </For>
-                <hr />
-                Layers; <br />
-                <button onClick={() => pushUILayer({
-                    id: String(Math.random()),
-                    component: () => TestComp(),
-                    blockBehind: true,
-                    style: {background: "#000000aa"}
-                })}>
-                    Add blocking test layer
-                </button>
-                <button onClick={() => pushUILayer({
-                    id: String(Math.random()),
-                    component: () => TestComp(),
-                })}>
-                    Add nonblocking test layer
-                </button>
-                <hr/>
-                Battles; <br />
-                <button onClick={() => startBattle(OPPONENT_PANOPTES).then((outcome) => alert(`Battle outcome handled by .then on button: ${outcome}`))}>Panoptes</button>
-                <button onClick={() => startBattle(OPPONENT_DEBUG_ANGEL)}>Debug Angel</button>
-            </CornerRect>
-        </CornerRect>
-    )
-}
